% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.cpdist.R
\name{cpdist}
\alias{cpdist}
\title{Compare different distributions for their fits to the data.}
\usage{
cpdist(m1, m2, output = F)
}
\arguments{
\item{m1}{Fitted model on the left hand side of the function}

\item{m2}{Fitted model on the right hand side of the function}
}
\value{
If output = F, returns two-sided and one-side p-values. If output = T, returns the complete output of Vuong's test.
}
\description{
Compare different distributions for their fits to the data.
}
\examples{
cpdist(model.pl, model.ln, output = F)
cpdist(model.pl, model.ex, output = F)
# see vignette for explanations on how to interpret the p-values of Vuong's test.
}
