% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.fitex.R
\name{fitex}
\alias{fitex}
\title{Fit an exponential distribution to the degree distribution of the language network.}
\usage{
fitex(distribution, estimates = F)
}
\arguments{
\item{distribution}{A vector of integers representing the degree distribution of a given network.}
}
\value{
If estimates = F, returns the fitted model for subsequent analyses. If estimates = T, returns optimal xmin and parameters.
}
\description{
Fit an exponential distribution to the degree distribution of the language network.
}
\examples{
fitex(distribution, estimates = T)
# xmin, par1 values will be listed
model.ex <- fitln(distribution, estimates = F)
# use this model for other functions

}
