% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.fitln.R
\name{fitln}
\alias{fitln}
\title{Fit a log normal distribution to the degree distribution of the language network.}
\usage{
fitln(distribution, estimates = F)
}
\arguments{
\item{distribution}{A vector of integers representing the degree distribution of a given network.}
}
\value{
If estimates = F, returns the fitted model for subsequent analyses. If estimates = T, returns optimal xmin and parameters.
}
\description{
Fit a log normal distribution to the degree distribution of the language network.
}
\examples{
fitln(distribution, estimates = T)
# xmin, par1, par2 values will be listed
model.ln <- fitln(distribution, estimates = F)
# use this model for other functions

}
