% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.fitpl.R
\name{fitpl}
\alias{fitpl}
\title{Fit a power law distribution to the degree distribution of the language network.}
\usage{
fitpl(distribution, estimates = F)
}
\arguments{
\item{distribution}{A vector of integers representing the degree distribution of a given network.}
}
\value{
If estimates = F, returns the fitted model for subsequent analyses. If estimates = T, returns optimal xmin and parameters.
}
\description{
Fit a power law distribution to the degree distribution of the language network.
}
\examples{
fitpl(distribution, estimates = T)
# xmin, alpha (par1) values will be listed
model.pl <- fitpl(distribution, estimates = F)
# use this model for other functions

}
