% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.fitpo.R
\name{fitpo}
\alias{fitpo}
\title{Fit a Poisson distribution to the degree distribution of the language network.}
\usage{
fitpo(distribution, estimates = F)
}
\arguments{
\item{distribution}{A vector of integers representing the degree distribution of a given network.}
}
\value{
If estimates = F, returns the fitted model for subsequent analyses. If estimates = T, returns optimal xmin and parameters.
}
\description{
Fit a Poisson distribution to the degree distribution of the language network.
}
\examples{
fitpo(distribution, estimates = T)
# xmin, par1 values will be listed
model.po <- fitln(distribution, estimates = F)
# use this model for other functions

}
