% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.getdegdist.R
\name{getdegdist}
\alias{getdegdist}
\title{Fit a power law distribution to the degree distribution of the language network.}
\usage{
getdegdist(network)
}
\arguments{
\item{network}{The language network. Must be an igraph object.}
}
\value{
Degree distribution of network.
}
\description{
Fit a power law distribution to the degree distribution of the language network.
}
\examples{
network <- watts.strogatz.game(1, 5000, 1, 0.35, loops = FALSE, multiple = FALSE) # a fake network
distribution <- getdegdist(network)
fitpl(distribution) # fit a power law to the distribution
fitln(distribution) # fit a log normal distribution instead

}
