% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.getplsd.R
\name{getplsd}
\alias{getplsd}
\title{Get uncertainty estimates for power law parameters fitted to the degree distribution of the language network.}
\usage{
getplsd(model, getraw = F)
}
\arguments{
\item{model}{A power law model already fitted to the raw data.}
}
\value{
If getraw = F, returns the raw bootstrap output. Can be used for further analysis or visualization. If getraw = T, returns means and sds of the estimated parameters from the bootstrap.
}
\description{
Get uncertainty estimates for power law parameters fitted to the degree distribution of the language network.
}
\examples{
getplsd(model.pl, getraw = F)
# xmin mean, xmin sd, alpha mean, alpha sd values will be listed
# these are the means and sds of the simulations, not the actual data
# if the fitted alpha fits falls within the bootstrap distribution, then that's good
bsresults <- getplsd(model.pl, getraw = T)
# returns the raw outputs of the bootstrap

}
