% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giantc.R
\name{giantc}
\alias{giantc}
\title{Extract list of nodes in the giant component of the language network.}
\usage{
giantc(langnet)
}
\arguments{
\item{langnet}{An igraph object created using the \code{tolangnet} function and with label names created with the \code{nodeindex} function.}
}
\value{
A list of nodes in the LCC of the language network.
}
\description{
Extract list of nodes in the giant component of the language network.
}
\examples{
somewords <- c('cat', 'bat', 'cap', 'cape')
somewordsnet <- tolangnet(somewords)
somewordsnet <- nodeindex(somewordsnet, somewords)
somewords.gc <- giantc(somewordsnet)
somewords.gc

}
