% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermits.R
\name{hermits}
\alias{hermits}
\title{Extract list of hermits from an igraph network object.}
\usage{
hermits(langnet)
}
\arguments{
\item{langnet}{An igraph object created using the \code{tolangnet} function and with label names created with the \code{nodeindex} function.}
}
\value{
A list of hermits (nodes with no neighbors).
}
\description{
Extract list of hermits from an igraph network object.
}
\examples{
somewords <- c('cat', 'bat', 'cap', 'cape')
somewordsnet <- tolangnet(somewords)
somewordsnet <- nodeindex(somewordsnet, somewords)
somewords.hermits <- hermits(somewordsnet)
hermits # to view hermits
}
