% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeindex.R
\name{nodeindex}
\alias{nodeindex}
\title{Replaces node IDs with their labels.}
\usage{
nodeindex(langnet, wordlist)
}
\arguments{
\item{langnet}{An igraph object created using the \code{tolangnet} function.}

\item{wordlist}{A list of words. Must be a character vector.}
}
\value{
An igraph object of the language network with updated labels.
}
\description{
Replaces node IDs with their labels.
}
\examples{
somewords <- c('cat', 'bat', 'cap', 'cape')
somewordsnet <- tolangnet(somewords)
plot(somewordsnet) #plots the graph

somewordsnet.labelled <- nodeindex(somewordsnet, somewords)
plot(somewordsnet.labelled) #notice that this network has labels...
# note that the generated IDs are dependent on the ordering in the \\code{wordlist}.
}
