% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.testpl.R
\name{testpl}
\alias{testpl}
\title{Test the fit of a power law to the degree distribution of the language network.}
\usage{
testpl(model, getraw = F)
}
\arguments{
\item{model}{A power law model already fitted to the raw data.}
}
\value{
If getraw = F, returns the raw bootstrap output. Can be used for further analysis or visualization. If getraw = T, returns means and sds of the estimated parameters from the bootstrap.
}
\description{
Test the fit of a power law to the degree distribution of the language network.
}
\examples{
testpl(model.pl, getraw = F)
# xmin mean, xmin sd, alpha mean, alpha sd, p value will be listed
# interpreting the p-value: if p value is not sig, cannot reject H0
# i.e., we cannot rule out the power law model
# H0: data is generated from a power law distribution.
# H1: data is not generated from a power law distribution.
bspresults <- testpl(model.pl, getraw = T)
# returns the raw outputs of the bootstrap

}
