% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toedgelist.R
\name{toedgelist}
\alias{toedgelist}
\title{Converts an igraph network object to an edgelist with labels.}
\usage{
toedgelist(langnet)
}
\arguments{
\item{langnet}{An igraph object created using the \code{tolangnet} function and with label names created with the \code{nodeindex} function.}
}
\value{
A dataframe of the *labeled* edgelist of the language network created from \code{wordlist}.
}
\description{
Converts an igraph network object to an edgelist with labels.
}
\examples{
somewords <- c('cat', 'bat', 'cap', 'cape')
somewordsnet <- tolangnet(somewords)
somewordsnet <- nodeindex(somewordsnet, somewords)
somewords.edgelist <- toedgelist(somewordsnet)
head(somewords.edgelist)
}
