% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize the fits of various distributions to the degree distribution of the language network.}
\usage{
visualize(distribution)
}
\arguments{
\item{distribution}{A vector of integers representing the degree distribution of a given network.}
}
\value{
A figure with the distribution fits to the raw data.
}
\description{
Visualize the fits of various distributions to the degree distribution of the language network.
}
\examples{
# visualize(distribution)

}
